Special _guard_result parameter on Guard'ed callables for custom behaviour if unauthorized
Normally, Forbidden (zExceptions) is raised if the Guard denies execution.
But we'd like to use protect scripts like Base_edit with Guards (so we can test
security without actually calling the script) and still return a nice response
to the user if the document became non-editable while she was filling the form.

Any ExternalMethod/PythonScript with a '_guard_result' argument is filled
automatically with a false value if the Guard check fails.

From the point of view of the caller, it's like the callable does not take
such parameter, and it has no way to pass it.

_guard_result can be placed anywhere (except before self in the case of
External Methods). The only allowed default value is None.